split DataBundle class
